replace sqrt, hypot, sin, cos with float variants (ie sqrtf, hypotf, sinf, cosf)

+++ src/modules/robot/Block.cpp
@@ -100,7 +100,7 @@ float Block::intersection_distance(float initialrate, float finalrate, float acc
 // acceleration within the allotted distance.
 inline float max_allowable_speed(float acceleration, float target_velocity, float distance) {
   return(
-    sqrt(target_velocity*target_velocity-2L*acceleration*distance)  //Was acceleration*60*60*distance, in case this breaks, but here we prefer to use seconds instead of minutes
+    sqrtf(target_velocity*target_velocity-2L*acceleration*distance)  //Was acceleration*60*60*distance, in case this breaks, but here we prefer to use seconds instead of minutes
   );
 }
